D2NOVA Stephen Chen
In latest Chrome on MacOS (Version 60.0.3080.5)
the angular-locker 2.0.4 will fail to read from localStorage.
i.e. locker.get(mykey) will return undefined even there is data stored with the key 'mykey'
the root cause is that in _exists() method

this._driver.hasOwnProperty(this._getPrefix(_value(key)))

is not working for this Chrome.

so the fix to that is to change to :

this._driver.hasOwnProperty(this._getPrefix(_value(key))) || this._getItem(key)